Add missing frontend tests
