+++ tests/Frontend/e2e/RoomsViewGeneral.cy.js
@@ -6,958 +6,6 @@ describe("Room View general", function () {
     cy.interceptRoomViewRequests();
   });
 
-  it("room view as guest", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    cy.fixture("room.json").then((room) => {
-      room.data.allow_membership = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access overlay is shown
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        // Check that header is shown
-        cy.contains("Meeting One").should("be.visible");
-        cy.contains("John Doe").should("be.visible");
-        cy.contains("rooms.index.room_component.never_started").should(
-          "be.visible",
-        );
-
-        // Check that login button is shown
-        cy.get('[data-test="room-login-as-user-button"]')
-          .should("be.visible")
-          .and("have.text", "auth.offer_login")
-          .and("have.attr", "href", "/login?redirect=/rooms/abc-def-123");
-
-        // Check that participant name input is shown
-        cy.get('[data-test="participant-name-field"]')
-          .should("be.visible")
-          .within(() => {
-            cy.contains("rooms.first_and_lastname").should("be.visible");
-            cy.get("#participant-name")
-              .should("be.visible")
-              .and("have.value", "");
-            cy.contains("rooms.remember_participant_name").should("be.visible");
-            cy.get("#remember-participant-name").and("not.be.checked");
-          });
-
-        // Check that access code input is not shown
-        cy.get('[data-test="access-code-field"]').should("not.exist");
-
-        // Enter guest name
-        cy.get("#participant-name").type("Max Doe");
-
-        cy.get('[data-test="room-login-button"]')
-          .should("have.text", "rooms.continue_as_guest")
-          .click();
-      });
-
-    // Check that room access overlay is hidden
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("John Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // ToDo check that welcome message is shown
-
-    // Check that buttons are shown correctly
-    cy.get('[data-test="reload-room-button"]').should("be.visible");
-    cy.get('[data-test="room-join-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-end-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-favorites-button"]').should("not.exist");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("not.exist");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Check that correct tab is shown
-    cy.contains("rooms.files.title").should("be.visible");
-
-    // Check if share button is hidden
-    cy.get('[data-test="room-share-button"]').should("not.exist");
-
-    // Test reloading the room
-    cy.fixture("room.json").then((room) => {
-      room.data.name = "Meeting Two";
-      room.data.owner.id = 2;
-      room.data.owner.name = "Max Doe";
-      room.data.last_meeting = {
-        start: "2023-08-21T08:18:28.000000Z",
-        end: null,
-      };
-      room.data.description = "<p>Test</p>";
-      room.data.access_code = null;
-      room.data.current_user = null;
-
-      const reloadRequest = interceptIndefinitely(
-        "GET",
-        "api/v1/rooms/abc-def-123",
-        {
-          statusCode: 200,
-          body: room,
-        },
-        "roomRequest",
-      );
-
-      // Trigger reload
-      cy.get('[data-test="reload-room-button"]').click();
-      cy.get('[data-test="reload-room-button"]')
-        .should("be.disabled")
-        .then(() => {
-          reloadRequest.sendResponse();
-        });
-    });
-
-    cy.title().should("eq", "Meeting Two - PILOS Test");
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting Two").should("be.visible");
-    cy.contains("Max Doe").should("be.visible");
-    cy.contains(
-      'rooms.index.room_component.running_since_{"date":"08/21/2023, 04:18"}',
-    ).should("be.visible");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("be.visible");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-  });
-
-  it("room view with access code", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access code overlay is shown correctly
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        // Check header is shown
-        cy.contains("Meeting One").should("be.visible");
-        cy.contains("Max Doe").should("be.visible");
-        cy.contains("rooms.index.room_component.never_started").should(
-          "be.visible",
-        );
-
-        // Check that login button is hidden
-        cy.get('[data-test="room-login-as-user-button"]').should("not.exist");
-
-        // Check that participant name input is hidden
-        cy.get('[data-test="participant-name-field"]').should("not.exist");
-
-        // Check that access code input is shown
-        cy.get('[data-test="access-code-field"]')
-          .should("be.visible")
-          .within(() => {
-            cy.contains("rooms.access_code").should("be.visible");
-            cy.get("#access-code").should("be.visible").and("have.value", "");
-          });
-
-        // Try to submit with correct access code
-        cy.get("#access-code").type("123456789");
-      });
-
-    const roomAuthRequest = interceptIndefinitely(
-      "POST",
-      "api/v1/rooms/abc-def-123/auth",
-      {
-        statusCode: 201,
-        body: {
-          data: {
-            id: "roomAuthToken",
-            type: 0,
-          },
-        },
-      },
-      "roomAuthRequest",
-    );
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    // Check loading
-    cy.get("[data-test='reload-room-button']").should("be.disabled");
-    cy.get('[data-test="room-login-button"]')
-      .should("be.disabled")
-      .then(() => {
-        roomAuthRequest.sendResponse();
-      });
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "123456789",
-        type: 0,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("Max Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check that buttons are shown correctly
-    cy.get('[data-test="reload-room-button"]').should("be.visible");
-    cy.get('[data-test="room-join-membership-button"]').should("be.visible");
-    cy.get('[data-test="room-end-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-favorites-button"]').should("be.visible");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("be.visible");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Check that the correct tab is shown
-    cy.contains("rooms.description.title").should("be.visible");
-
-    // Check if share button is hidden
-    cy.get('[data-test="room-share-button"]').should("not.exist");
-
-    // Reload and try with valid access code again
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.window().then((win) => {
-      win.sessionStorage.removeItem("roomAccessCode_abc-def-123");
-    });
-
-    cy.reload();
-
-    cy.wait("@roomRequest");
-
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        // Try to submit with correct access code
-        cy.get("#access-code").type("123456789");
-      });
-
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 200,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 0,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "123456789",
-        type: 0,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-
-    // Reload with invalid access code
-    const errorReloadRoomRequest = interceptIndefinitely(
-      "GET",
-      "api/v1/rooms/abc-def-123*",
-      {
-        statusCode: 401,
-        body: {
-          message: "invalid_auth_token",
-        },
-      },
-      "roomRequest",
-    );
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    // Intercept second request (reload room) and send response of the first request
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      })
-        .as("roomRequest")
-        .then(() => {
-          errorReloadRoomRequest.sendResponse();
-        });
-    });
-
-    // Check that access code header is set for the first request
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-    // Check that access code header is reset for the second request (reload room)
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.not.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    // Check if error message is shown
-    cy.checkToastMessage("rooms.flash.access_code_invalid");
-
-    cy.contains("rooms.flash.access_code_invalid").should("be.visible");
-
-    cy.get('[data-test="room-access-overlay"]').should("be.visible");
-
-    // Retry with valid access code but no access code needed anymore
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        // Try to submit with correct access code
-        cy.get("#access-code").type("123456789");
-      });
-
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 204,
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "123456789",
-        type: 0,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query.room_auth_token).to.be.undefined;
-      expect(interception.request.query.room_auth_token_type).to.be.undefined;
-    });
-
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-  });
-
-  it("room view with legacy access code", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.legacy_code = true;
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access code input is shown correctly
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        cy.contains("Meeting One").should("be.visible");
-        cy.contains("Max Doe").should("be.visible");
-        cy.contains("rooms.index.room_component.never_started").should(
-          "be.visible",
-        );
-
-        // Submit valid access code
-        cy.get("#access-code").type("012abc");
-      });
-
-    // Intercept room auth request
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 0,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "012abc",
-        type: 0,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("Max Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check that buttons are shown correctly
-    cy.get('[data-test="reload-room-button"]').should("be.visible");
-    cy.get('[data-test="room-join-membership-button"]').should("be.visible");
-    cy.get('[data-test="room-end-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-favorites-button"]').should("be.visible");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("be.visible");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Check that the correct tab is shown
-    cy.contains("rooms.description.title").should("be.visible");
-  });
-
-  it("room auth with access code errors", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access overlay is shown correctly
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        cy.contains("Meeting One").should("be.visible");
-        cy.contains("Max Doe").should("be.visible");
-        cy.contains("rooms.index.room_component.never_started").should(
-          "be.visible",
-        );
-        // Try to submit without access code
-      });
-
-    // Check with 422 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 422,
-      body: {
-        message: "The Access code field is required.",
-        errors: {
-          access_code: ["The Access code field is required."],
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        cy.contains("The Access code field is required.").should("be.visible");
-
-        // Try to submit with invalid access code
-        cy.get("#access-code").type("987654321");
-      });
-
-    // Check with invalid_code error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 401,
-      body: {
-        message: "invalid_code",
-      },
-    }).as("roomAuthRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    // Intercept room request (reload room)
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    // Wait for room auth request and check if access code is set
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "987654321",
-        type: 0,
-      });
-    });
-
-    // Wait for room request
-    cy.wait("@roomRequest");
-
-    // Check if error message is shown
-    cy.checkToastMessage("rooms.flash.access_code_invalid");
-
-    cy.contains("rooms.flash.access_code_invalid").should("be.visible");
-
-    // Intercept room auth request and respond with rate limit error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 429,
-      body: {
-        limit: "room_auth",
-        retry_after: 5,
-      },
-    }).as("roomAuthRequest");
-
-    cy.clock();
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    // Wait for room auth request
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "987654321",
-        type: 0,
-      });
-    });
-
-    // Check if input and buttons are disabled
-    cy.get("#access-code").should("be.disabled");
-    cy.get('[data-test="room-login-button"]').should("be.disabled");
-    cy.get('[data-test="reload-room-button"]').should("be.disabled");
-
-    // Check countdown
-    for (let i = 5; i > 0; i--) {
-      // Check if countdown message is updated
-      cy.contains('rooms.auth_throttled_{"try_again":' + i + "}").should(
-        "be.visible",
-      );
-
-      // Tick clock 1 sec forward
-      cy.tick(1000);
-    }
-
-    // restore the clock
-    cy.clock().then((clock) => {
-      clock.restore();
-    });
-
-    // Check toast message
-    cy.checkToastMessage("app.flash.too_many_requests");
-
-    // Check if input and buttons are enabled again
-    cy.get("#access-code").should("not.be.disabled");
-    cy.get('[data-test="room-login-button"]').should("not.be.disabled");
-    cy.get('[data-test="reload-room-button"]').should("not.be.disabled");
-
-    // Check with 500 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 500,
-      body: {
-        message: "Test",
-      },
-    }).as("roomAuthRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.server_error.message_{"message":"Test"}',
-      'app.flash.server_error.error_code_{"statusCode":500}',
-    ]);
-
-    // Check that access overlay is still shown and not disabled
-    cy.get('[data-test="room-access-overlay"]').should("be.visible");
-    cy.get("#access-code").should("not.be.disabled");
-    cy.get('[data-test="room-login-button"]').should("not.be.disabled");
-
-    // Check with guests not allowed
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 403,
-      body: {
-        message: "guests_not_allowed",
-      },
-    }).as("roomAuthRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    // Check if error message is shown
-    // Check that the error message is shown
-    cy.contains("rooms.only_used_by_authenticated_users").should("be.visible");
-
-    // Check that access overlay is hidden
-    cy.get('[data-test="room-access-overlay"]').should("not.exist");
-
-    // Check with 404 error (room not found) as authenticated user
-    cy.interceptRoomIndexRequests();
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 404,
-      body: {
-        message: "model_not_found",
-        model: "room",
-        ids: ["abc-def-123"],
-      },
-    }).as("roomAuthRequest");
-
-    cy.reload();
-
-    cy.wait("@roomRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    // Check that redirect to room index page worked
-    cy.url()
-      .should("include", "/rooms")
-      .and("not.include", "/rooms/abc-def-123");
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.model_not_found.title_{"model":"app.model.room"}',
-      'app.flash.model_not_found.details_{"ids":"abc-def-123"}',
-    ]);
-
-    // Check with 404 error (room not found) as guest
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.fixture("room.json").then((room) => {
-      room.data.current_user = null;
-      room.data.authenticated = false;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.get("#participant-name").type("Max Doe");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    // Check that redirect to 404 page worked
-    cy.url().should("include", "/404").and("not.include", "/rooms/abc-def-123");
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.model_not_found.title_{"model":"app.model.room"}',
-      'app.flash.model_not_found.details_{"ids":"abc-def-123"}',
-    ]);
-  });
-
-  it("room view with access code errors", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access code input is shown correctly
-    cy.get("#access-code").type("123456789");
-
-    // Check with invalid token error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 0,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    const roomRequest = interceptIndefinitely(
-      "GET",
-      "api/v1/rooms/abc-def-123*",
-      {
-        statusCode: 401,
-        body: {
-          message: "invalid_auth_token",
-        },
-      },
-      "roomRequest",
-    );
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    // Wait for room auth request and check if access code is set
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "123456789",
-        type: 0,
-      });
-    });
-
-    cy.fixture("room.json")
-      .then((room) => {
-        room.data.owner = {
-          id: 2,
-          name: "Max Doe",
-        };
-        room.data.authenticated = false;
-        room.data.description = "<p>Test</p>";
-        room.data.allow_membership = true;
-
-        cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-          statusCode: 200,
-          body: room,
-        }).as("roomRequest");
-      })
-      .then(() => {
-        roomRequest.sendResponse();
-      });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query.room_auth_token).to.be.undefined;
-      expect(interception.request.query.room_auth_token_type).to.be.undefined;
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.access_code_invalid");
-
-    cy.contains("rooms.flash.access_code_invalid").should("be.visible");
-
-    cy.get('[data-test="room-access-overlay"]').should("be.visible");
-
-    // Check with 500 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 0,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 500,
-      body: {
-        message: "Test",
-      },
-    }).as("roomRequest");
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        access_code: "123456789",
-        type: 0,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.server_error.message_{"message":"Test"}',
-      'app.flash.server_error.error_code_{"statusCode":500}',
-    ]);
-
-    // Check that access code overlay is still shown and not disabled
-    cy.get('[data-test="room-access-overlay"]').should("be.visible");
-    cy.get("#access-code").should("not.be.disabled");
-    cy.get('[data-test="room-login-button"]').should("not.be.disabled");
-  });
-
   it("room view as member", function () {
     cy.interceptRoomFilesRequest();
     cy.fixture("room.json").then((room) => {
@@ -988,6 +36,11 @@ describe("Room View general", function () {
     cy.contains(
       'rooms.index.room_component.last_ran_till_{"date":"08/21/2023, 10:20"}',
     ).should("be.visible");
+
+    // Check that participant name input is hidden
+    cy.get('[data-test="participant-name-field"]').should("not.exist");
+    cy.contains("rooms.name_in_video_conference").should("not.exist");
+    cy.get('[data-test="change-participant-name-button"]').should("not.exist");
 
     // Check that buttons are shown correctly
     cy.get('[data-test="reload-room-button"]').should("be.visible");
@@ -1039,6 +92,11 @@ describe("Room View general", function () {
     cy.contains("rooms.index.room_component.never_started").should(
       "be.visible",
     );
+
+    // Check that participant name input is hidden
+    cy.get('[data-test="participant-name-field"]').should("not.exist");
+    cy.contains("rooms.name_in_video_conference").should("not.exist");
+    cy.get('[data-test="change-participant-name-button"]').should("not.exist");
 
     // Check that buttons are shown correctly
     cy.get('[data-test="reload-room-button"]').should("be.visible");
@@ -1302,6 +360,10 @@ describe("Room View general", function () {
       'rooms.index.room_component.running_since_{"date":"08/21/2023, 10:18"}',
     ).should("be.visible");
 
+    // Check that participant name is hidden
+    cy.contains("rooms.name_in_video_conference").should("not.exist");
+    cy.get('[data-test="change-participant-name-button"]').should("not.exist");
+
     // Check that buttons are shown correctly
     cy.get('[data-test="reload-room-button"]').should("be.visible");
     cy.get('[data-test="room-join-membership-button"]').should("not.exist");
@@ -1347,6 +409,10 @@ describe("Room View general", function () {
       "be.visible",
     );
 
+    // Check that participant name is hidden
+    cy.contains("rooms.name_in_video_conference").should("not.exist");
+    cy.get('[data-test="change-participant-name-button"]').should("not.exist");
+
     // Check that buttons are shown correctly
     cy.get('[data-test="reload-room-button"]').should("be.visible");
     cy.get('[data-test="room-join-membership-button"]').should("not.exist");
@@ -1367,583 +433,6 @@ describe("Room View general", function () {
 
     // Check if share button is shown correctly
     cy.get('[data-test="room-share-button"]').should("exist");
-  });
-
-  it("room view with personalized link (participant)", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    // Intercept room auth request
-    const roomAuthRequest = interceptIndefinitely(
-      "POST",
-      "api/v1/rooms/abc-def-123/auth",
-      {
-        statusCode: 201,
-        body: {
-          data: {
-            id: "roomAuthToken",
-            type: 1,
-          },
-        },
-      },
-      "roomAuthRequest",
-    );
-
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123#personalizedLink=xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.get("[data-test='room-loading-spinner']")
-      .should("be.visible")
-      .then(() => {
-        roomAuthRequest.sendResponse();
-      });
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that header for token is set
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("John Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check that buttons are shown correctly
-    cy.get('[data-test="reload-room-button"]').should("be.visible");
-    cy.get('[data-test="room-join-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-end-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-favorites-button"]').should("not.exist");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("not.exist");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Check that correct tab is shown
-    cy.contains("rooms.files.title").should("be.visible");
-
-    // Check if share button is hidden
-    cy.get('[data-test="room-share-button"]').should("not.exist");
-
-    // Reload with invalid token
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 401,
-      body: {
-        message: "invalid_personalized_link",
-      },
-    }).as("roomAuthRequest");
-
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 401,
-      body: {
-        message: "invalid_auth_token",
-      },
-    }).as("roomRequest");
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.personalized_link_invalid");
-    cy.contains("rooms.invalid_personalized_link").should("be.visible");
-  });
-
-  it("room view with personalized link (moderator)", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    // Intercept room auth request
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.is_moderator = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123#personalizedLink=xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that header for token is set
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    // Check that room Header is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("John Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check that buttons are shown correctly
-    cy.get('[data-test="reload-room-button"]').should("be.visible");
-    cy.get('[data-test="room-join-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-end-membership-button"]').should("not.exist");
-    cy.get('[data-test="room-favorites-button"]').should("not.exist");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("not.exist");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Check that correct tab is shown
-    cy.contains("rooms.files.title").should("be.visible");
-
-    // Check if share button is hidden
-    cy.get('[data-test="room-share-button"]').should("not.exist");
-  });
-
-  it("room auth with personalized link errors", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    // 401 invalid personalize link
-    // Intercept room auth request
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 401,
-      body: {
-        message: "invalid_personalized_link",
-      },
-    }).as("roomAuthRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.personalized_link_invalid");
-    cy.contains("rooms.invalid_personalized_link").should("be.visible");
-
-    // Reload and check with 422 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 422,
-      body: {
-        message: "The Access token field is required.",
-        errors: {
-          personalized_link_token: ["The Access token field is required."],
-        },
-      },
-    }).as("roomAuthRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.personalized_link_invalid");
-    cy.contains("rooms.invalid_personalized_link").should("be.visible");
-
-    // Check with guests only error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 420,
-      body: {
-        message: "guests_only",
-      },
-    }).as("roomAuthRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    cy.checkToastMessage("app.flash.guests_only");
-    cy.url()
-      .should("not.include", "/rooms")
-      .and("not.include", "rooms/abc-def-123");
-
-    // Check with 500 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 500,
-      body: {
-        message: "Test",
-      },
-    }).as("roomAuthRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.server_error.message_{"message":"Test"}',
-      'app.flash.server_error.error_code_{"statusCode":500}',
-    ]);
-
-    // Check that reload button is shown
-    cy.get('[data-test="reload-button"]').should("be.visible");
-
-    // Click reload button and make sure auth request is sent again
-    // Reload with valid auth request and room request
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="reload-button"]').click();
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    // Check that room is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("John Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check with 404 error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 404,
-      body: {
-        message: "model_not_found",
-        model: "room",
-        ids: ["abc-def-123"],
-      },
-    }).as("roomAuthRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that redirect to 404 page worked and error message is shown
-    cy.url().should("include", "/404").and("not.include", "rooms/abc-def-123");
-
-    cy.checkToastMessage([
-      'app.flash.model_not_found.title_{"model":"app.model.room"}',
-      'app.flash.model_not_found.details_{"ids":"abc-def-123"}',
-    ]);
-  });
-
-  it("room view with personalized link errors", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    // Check with 401 invalid token error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    const roomRequest = interceptIndefinitely(
-      "GET",
-      "api/v1/rooms/abc-def-123*",
-      {
-        statusCode: 401,
-        body: {
-          message: "invalid_auth_token",
-        },
-      },
-      "roomRequest",
-    );
-
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-
-      cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-        statusCode: 401,
-        body: {
-          message: "invalid_personalized_link",
-        },
-      }).as("roomAuthRequest");
-
-      roomRequest.sendResponse();
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.personalized_link_invalid");
-    cy.contains("rooms.invalid_personalized_link").should("be.visible");
-
-    // Check with guests only error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 420,
-      body: {
-        message: "guests_only",
-      },
-    }).as("roomRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    cy.checkToastMessage("app.flash.guests_only");
-    cy.url()
-      .should("not.include", "/rooms")
-      .and("not.include", "rooms/abc-def-123");
-
-    // Check with 500 error
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 500,
-      body: {
-        message: "Test",
-      },
-    }).as("roomRequest");
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest").then((interception) => {
-      expect(interception.request.body).to.eql({
-        personalized_link_token:
-          "xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-        type: 1,
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage([
-      'app.flash.server_error.message_{"message":"Test"}',
-      'app.flash.server_error.error_code_{"statusCode":500}',
-    ]);
-
-    // Check that reload button is shown
-    cy.get('[data-test="reload-button"]').should("be.visible");
-
-    // Intercept room auth request with different token
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "differentToken",
-          type: 1,
-        },
-      },
-    }).as("differentRoomAuthRequest");
-
-    // Click reload button and make sure room request is sent again with same auth token
-    // Reload with valid room request
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="reload-button"]').click();
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    //  Check that room is shown correctly
-    cy.contains("Meeting One").should("be.visible");
-    cy.contains("John Doe").should("be.visible");
-    cy.contains("rooms.index.room_component.never_started").should(
-      "be.visible",
-    );
-
-    // Check that room auth request was not sent again
-    cy.get("@differentRoomAuthRequest").should("be.null");
   });
 
   it("room view with rooms.viewAll permission", function () {
@@ -2148,7 +637,6 @@ describe("Room View general", function () {
     cy.get('[data-test="dialog-close-button"]').click();
   });
 
-  // ToDo debug this and check why access code does not stay if visited with access code
   it("membership button", function () {
     cy.fixture("room.json").then((room) => {
       room.data.owner = {
@@ -3093,35 +1581,6 @@ describe("Room View general", function () {
     );
   });
 
-  it("visit with personalized link as authenticated user", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.is_member = true;
-      room.data.is_moderator = true;
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    // Visit room with personalized link
-    cy.visit(
-      "/rooms/abc-def-123#personalizedLink=xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    // Check that error message is shown and user is redirected to the home page
-    cy.checkToastMessage("app.flash.guests_only");
-    cy.url().should(
-      "not.include",
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-  });
-
   it("visit with general error", function () {
     cy.intercept("GET", "api/v1/rooms/abc-def-123", {
       statusCode: 500,
@@ -3376,322 +1835,6 @@ describe("Room View general", function () {
       'app.flash.model_not_found.title_{"model":"app.model.room"}',
       'app.flash.model_not_found.details_{"ids":"abc-def-123"}',
     ]);
-  });
-
-  it("reload with access code errors", function () {
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    cy.wait("@roomRequest");
-
-    cy.title().should("eq", "Meeting One - PILOS Test");
-
-    // Check that access code input is shown correctly
-    cy.get("#access-code").type("123456789");
-
-    // Check with invalid token error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 0,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = true;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="room-login-button"]').click();
-
-    cy.wait("@roomAuthRequest");
-
-    cy.wait("@roomRequest");
-
-    // Test reload with invalid token error
-
-    const roomRequest = interceptIndefinitely(
-      "GET",
-      "api/v1/rooms/abc-def-123*",
-      {
-        statusCode: 401,
-        body: {
-          message: "invalid_auth_token",
-        },
-      },
-      "roomRequest",
-    );
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    cy.fixture("room.json").then((room) => {
-      room.data.owner = {
-        id: 2,
-        name: "Max Doe",
-      };
-      room.data.authenticated = false;
-      room.data.description = "<p>Test</p>";
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      })
-        .as("roomRequest")
-        .then(() => {
-          roomRequest.sendResponse();
-        });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "0",
-      });
-    });
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query.room_auth_token).to.be.undefined;
-      expect(interception.request.query.room_auth_token_type).to.be.undefined;
-    });
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.access_code_invalid");
-
-    cy.contains("rooms.flash.access_code_invalid").should("be.visible");
-
-    cy.get('[data-test="room-access-overlay"]').should("be.visible");
-  });
-
-  it("reload with personalized link errors", function () {
-    cy.intercept("GET", "api/v1/currentUser", {});
-    cy.interceptRoomFilesRequest();
-
-    // Check with 401 invalid token error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit(
-      "/rooms/abc-def-123/xWDCevVTcMys1ftzt3nFPgU56Wf32fopFWgAEBtklSkFU22z1ntA4fBHsHeMygMiOa9szJbNEfBAgEWSLNWg2gcF65PwPZ2ylPQR",
-    );
-
-    cy.wait("@roomAuthRequest");
-    cy.wait("@roomRequest");
-
-    // Check with invalid token error
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 401,
-      body: {
-        message: "invalid_auth_token",
-      },
-    }).as("roomRequest");
-
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 401,
-      body: {
-        message: "invalid_personalized_link",
-      },
-    }).as("roomAuthRequest");
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    cy.wait("@roomRequest").then((interception) => {
-      expect(interception.request.query).to.contain({
-        room_auth_token: "roomAuthToken",
-        room_auth_token_type: "1",
-      });
-    });
-
-    cy.wait("@roomAuthRequest");
-
-    // Check that error message is shown
-    cy.checkToastMessage("rooms.flash.personalized_link_invalid");
-    cy.contains("rooms.invalid_personalized_link").should("be.visible");
-
-    // Check with guests only error
-    cy.intercept("POST", "api/v1/rooms/abc-def-123/auth", {
-      statusCode: 201,
-      body: {
-        data: {
-          id: "roomAuthToken",
-          type: 1,
-        },
-      },
-    }).as("roomAuthRequest");
-
-    cy.fixture("room.json").then((room) => {
-      room.data.username = "Max Doe";
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.reload();
-
-    cy.wait("@roomAuthRequest");
-    cy.wait("@roomRequest");
-
-    cy.intercept("GET", "api/v1/rooms/abc-def-123*", {
-      statusCode: 420,
-      body: {
-        message: "guests_only",
-      },
-    }).as("roomRequest");
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    cy.wait("@roomRequest");
-
-    // Check that error message is shown
-    cy.checkToastMessage("app.flash.guests_only");
-    cy.url()
-      .should("not.include", "/rooms")
-      .and("not.include", "rooms/abc-def-123");
-  });
-
-  it("logged in status change", function () {
-    cy.interceptRoomFilesRequest();
-    cy.fixture("room.json").then((room) => {
-      room.data.allow_membership = true;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.visit("/rooms/abc-def-123");
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("be.visible");
-    cy.get("#tab-members").should("be.visible");
-    cy.get("#tab-tokens").should("be.visible");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("be.visible");
-    cy.get("#tab-settings").should("be.visible");
-
-    // Change current user to guest
-    cy.fixture("room.json").then((room) => {
-      room.data.allow_membership = true;
-      room.data.current_user = null;
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    // Check that access overlay is shown
-    cy.get('[data-test="room-access-overlay"]')
-      .should("be.visible")
-      .within(() => {
-        // Check that login button is shown
-        cy.get('[data-test="room-login-as-user-button"]').should("be.visible");
-
-        // Check that guest name input is shown and enter name
-        cy.get("#participant-name").should("be.visible").type("Max Doe");
-
-        // Check that access code input is hidden
-        cy.get('[data-test="access-code-field"]').should("not.exist");
-
-        // Login
-        cy.get('[data-test="room-login-button"]').click();
-      });
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("not.exist");
-    cy.get("#tab-members").should("not.exist");
-    cy.get("#tab-tokens").should("not.exist");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("not.exist");
-    cy.get("#tab-settings").should("not.exist");
-
-    // Change current user to co_owner
-    cy.fixture("room.json").then((room) => {
-      room.data.allow_membership = true;
-      room.data.is_member = true;
-      room.data.is_co_owner = true;
-      room.data.current_user = {
-        id: 2,
-        firstname: "Max",
-        lastname: "Doe",
-        user_locale: "en",
-        permissions: ["rooms.create"],
-        model_name: "User",
-        room_limit: -1,
-      };
-
-      cy.intercept("GET", "api/v1/rooms/abc-def-123", {
-        statusCode: 200,
-        body: room,
-      }).as("roomRequest");
-    });
-
-    cy.get('[data-test="reload-room-button"]').click();
-
-    // Check that tabs are shown correctly
-    cy.get("#tab-description").should("be.visible");
-    cy.get("#tab-members").should("be.visible");
-    cy.get("#tab-tokens").should("be.visible");
-    cy.get("#tab-files").should("be.visible");
-    cy.get("#tab-recordings").should("be.visible");
-    cy.get("#tab-history").should("be.visible");
-    cy.get("#tab-settings").should("be.visible");
   });
 
   it("displays meeting ended reason", function () {
